refactor: LoggerMiddleware 코드분리
