+++ src/app.module.ts
@@ -1,12 +1,4 @@
-import {
-  Injectable,
-  Logger,
-  MiddlewareConsumer,
-  Module,
-  NestMiddleware,
-  NestModule,
-} from '@nestjs/common';
-import { NextFunction, Request, Response } from 'express';
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { TypeOrmConfigService } from './config/typeorm.config';
@@ -20,23 +12,7 @@ import { SseModule } from './sse/sse.module';
 import { EventsModule } from './events/evnets.module';
 import { MessageService } from './message/message.service';
 import { MessageModule } from './message/message.module';
-
-@Injectable()
-export class LoggerMiddleware implements NestMiddleware {
-  private logger = new Logger('HTTP');
-
-  use(req: Request, res: Response, next: NextFunction) {
-    const { method, originalUrl } = req;
-    const userAgent = req.get('user-agent') || '';
-    res.on('finish', () => {
-      const { statusCode } = res;
-      this.logger.log(
-        `${method} ${statusCode} - ${originalUrl} - ${userAgent}`,
-      );
-    });
-    next();
-  }
-}
+import { LoggerMiddleware } from './global/middleware/logger.middleware';
 
 @Module({
   imports: [
